enhance repository consolidation script
- add failsafe to repo consolidation script
- stub in a solution for working w/ shallow repository

+++ gradle/consolidate-publish-repo.sh
@@ -6,8 +6,8 @@
 
 DRY_RUN=false
 QUIET_FLAG=
-MAX_AGE='1 month'
-FREQUENCY='1 week'
+MAX_AGE='3 week'
+FREQUENCY='1 hour'
 PUBLISH_DIR=_publish
 
 while getopts 'dqa:f:D:' option; do
@@ -21,6 +21,18 @@ while getopts 'dqa:f:D:' option; do
 done
 
 cd $PUBLISH_DIR
+
+# NOTE using the GitHub API would allow us to start with a shallow repository
+# UNTIL_PARAM_VAL="$MAX_AGE $FREQUENCY"
+# UNTIL_PARAM_VAL=${UNTIL_PARAM_VAL// /%20}
+# if [ "0" == `curl -s https://api.github.com/repos/mulesoft/mulesoft-docs-build-output/commits?until=$UNTIL_PARAM_VAL -H "Accept: application/vnd.github.json" | grep -c sha` ]; then
+#   echo 'No commits to consolidate.'
+#   exit 0
+# fi
+#
+# if [ -f .git/shallow ]; then
+#   git fetch --unshallow
+# fi
 
 # use schmitt trigger so script only runs every so often
 if [ -z `git log -n1 --format=%H --until "$MAX_AGE $FREQUENCY"` ]; then
@@ -55,8 +67,10 @@ git reflog expire --expire=now --all
 git gc $QUIET_FLAG --prune=now
 if [ "$DRY_RUN" == "false" ]; then
   git push $QUIET_FLAG --force origin master
-  cd ..
-  rm -rf $PUBLISH_DIR
+  if [ $? -ne 0 ]; then
+    cd ..
+    rm -rf $PUBLISH_DIR
+  fi
 fi
 
 echo 'Consolidation complete.'
